+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Change log
 
 ### Not yet released
+- Fix issue with size_t deserialization
 - Free functions version with `high` and `low` as template parameters
 - Free functions for bits insertion / extraction
 - Implementation's detail hidden under 'bits::detail' namespace